pyproject.toml
[project]
name = "prompter"
version = "0.1.0"
requires-python = ">= 3.12"
dependencies = [
  "mypy>=1.14.1",
  "openai>=1.60.2",
  "python-dotenv>=1.0.1",
  "ruff>=0.9.4",
]

[tool.mypy]
exclude = 'files/'
